refactor(core): split group service, controller, utils files
