+++ packages/canvas-engine/document/src/services/flow-group-service/index.ts
@@ -1,0 +1,2 @@
+export { FlowGroupController } from './flow-group-controller';
+export { FlowGroupService } from './flow-group-service';
